third_party/blink/web_tests/virtual/document-pip/document-pip-requires-user-gesture.html

<!DOCTYPE html>
<title>Test that navigator.documentPictureInPicture.requestWindow()
  fails without a user gesture</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
promise_test(t => {
  return promise_rejects_dom(t, 'NotAllowedError',
    navigator.documentPictureInPicture.requestWindow());
},
    "requestWindow should fail without a user gesture");
</script>
</body>
